# legacy/alert/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Alert Agent API",
        "description": "HTTP API for alert submission, prompt management, and WebSocket real-time alert broadcasting",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Alert Verification microservice endpoint"
        }
    ],
    "paths": {
...
        "/api/v1/realtime/incidents": {
            "get": {
                "tags": [
                    "realtime"
                ],
                "summary": "List incidents from Elasticsearch",
                "description": "Query incidents from Elasticsearch with optional filtering by sensor_id, category, and time range. Supports pagination via limit and offset.",
                "operationId": "list_incidents_api_v1_realtime_incidents_get",
                "parameters": [
                    {
                        "name": "sensor_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Filter by sensor ID",
                            "title": "Sensor Id"
                        },
                        "description": "Filter by sensor ID"
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Filter by incident category",
                            "title": "Category"
                        },
                        "description": "Filter by incident category"
                    },
                    {
                        "name": "start_time",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string",
                                    "format": "date-time"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Filter incidents after this ISO-8601 timestamp (e.g. 2024-01-15T10:30:00Z)",
                            "title": "Start Time"
                        },
                        "description": "Filter incidents after this ISO-8601 timestamp (e.g. 2024-01-15T10:30:00Z)"
                    },
                    {
                        "name": "end_time",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string",
                                    "format": "date-time"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Filter incidents before this ISO-8601 timestamp (e.g. 2024-01-15T18:00:00Z)",
                            "title": "End Time"
                        },
                        "description": "Filter incidents before this ISO-8601 timestamp (e.g. 2024-01-15T18:00:00Z)"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 1000,
                            "minimum": 1,
                            "description": "Maximum number of incidents to return",
                            "default": 100,
                            "title": "Limit"
                        },
                        "description": "Maximum number of incidents to return"
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 0,
                            "description": "Number of incidents to skip (for pagination)",
                            "default": 0,
                            "title": "Offset"
                        },
                        "description": "Number of incidents to skip (for pagination)"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Incidents list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IncidentListResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid timestamp format",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RealtimeAlertErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Elasticsearch query failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RealtimeAlertErrorResponse"
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "Elasticsearch unavailable",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RealtimeAlertErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/incidents": {
            "post": {
                "tags": [
                    "incident-submission"
                ],
                "summary": "Submit Incident for Processing",
                "description": "Submit a new incident (NvSchema) for processing and publish to Kafka. Accepts NvSchema Incident JSON (default) or a serialized Protobuf Incident (set Content-Type: application/x-protobuf).",
                "operationId": "submit_incident_api_v1_incidents_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "string"
                                    },
                                    "timestamp": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "end": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "sensorId": {
                                        "type": "string"
                                    },
                                    "category": {
                                        "type": "string"
                                    },
                                    "info": {
                                        "type": "object"
                                    },
                                    "event": {
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "required": [
                                    "id",
                                    "timestamp",
                                    "sensorId"
                                ]
                            },
                            "example": {
                                "id": "incident-67890",
                                "timestamp": "2025-01-15T14:30:00Z",
                                "end": "2025-01-15T14:30:30Z",
                                "sensorId": "cam_warehouse_02",
                                "category": "collision",
                                "info": {
                                    "media_urls": [
                                        "http://localhost:30888/vst/sim/media/incident.mp4"
                                    ],
                                    "media_type": "video"
                                },
                                "event": {
                                    "id": "incident-67890",
                                    "type": "incident_event"
                                }
                            }
                        },
                        "application/x-protobuf": {
                            "schema": {
                                "type": "string",
                                "format": "binary"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "202": {
                        "description": "Incident accepted and queued",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AlertSubmissionResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid Protobuf payload",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid JSON body",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AlertSubmissionResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "maxLength": 32,
                        "title": "Status",
                        "description": "Submission status"
                    },
                    "id": {
                        "type": "string",
                        "maxLength": 256,
                        "title": "Id",
                        "description": "Original event ID (used for correlation)"
                    },
                    "message": {
                        "type": "string",
                        "maxLength": 2000,
                        "title": "Message",
                        "description": "Human-readable status message"
                    },
                    "timestamp": {
                        "type": "string",
                        "maxLength": 64,
                        "title": "Timestamp",
                        "description": "Processing timestamp (ISO 8601)"
                    }
                },
                "type": "object",
                "required": [
                    "status",
                    "id",
                    "message",
                    "timestamp"
                ],
                "title": "AlertSubmissionResponse",
                "description": "HTTP response schema for successful alert submission.",
                "example": {
                    "id": "evt-12345-67890",
                    "message": "Alert queued for processing",
                    "status": "accepted",
                    "timestamp": "2025-01-15T14:30:05Z"
                }
            },
            "ErrorResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "maxLength": 32,
                        "title": "Status",
                        "description": "Error status"
                    },
                    "error": {
                        "type": "string",
                        "maxLength": 64,
                        "title": "Error",
                        "description": "Error type"
                    },
                    "message": {
                        "type": "string",
                        "maxLength": 2000,
                        "title": "Message",
                        "description": "Error message"
                    },
                    "timestamp": {
                        "type": "string",
                        "maxLength": 64,
                        "title": "Timestamp",
                        "description": "Error timestamp (ISO 8601)"
                    }
                },
                "type": "object",
                "required": [
                    "status",
                    "error",
                    "message",
                    "timestamp"
                ],
                "title": "ErrorResponse",
                "description": "HTTP error response schema.",
                "example": {
                    "error": "validation_failed",
                    "message": "Request validation failed with 3 error(s). Please check the request format and required fields.",
                    "status": "error",
                    "timestamp": "2025-01-15T14:30:05Z"
                }
            },
...
            "IncidentListResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "title": "Status",
                        "default": "success"
                    },
                    "incidents": {
                        "items": {
                            "additionalProperties": true,
                            "type": "object"
                        },
                        "type": "array",
                        "title": "Incidents",
                        "description": "List of incident documents from Elasticsearch",
                        "default": []
                    },
                    "count": {
                        "type": "integer",
                        "title": "Count",
                        "description": "Number of incidents returned"
                    },
                    "total": {
                        "type": "integer",
                        "title": "Total",
                        "description": "Total number of matching incidents in ES"
                    },
                    "timestamp": {
                        "type": "string",
                        "title": "Timestamp",
                        "description": "ISO-8601 response timestamp"
                    }
                },
                "type": "object",
                "required": [
                    "count",
                    "total",
                    "timestamp"
                ],
                "title": "IncidentListResponse",
                "description": "Response for GET /api/v1/realtime/incidents.",
                "example": {
                    "count": 1,
                    "incidents": [
                        {
                            "category": "collision",
                            "description": "Vehicle collision detected at intersection.",
                            "id": "incident-001",
                            "sensor_id": "cc06804c-7f11-4865-bb00-6b2db072086f",
                            "timestamp": "2025-06-01T12:00:00Z"
                        }
                    ],
                    "status": "success",
                    "timestamp": "2025-06-01T12:05:00Z",
                    "total": 42
                }
            },
...
            "RealtimeAlertErrorResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "title": "Status",
                        "default": "error"
                    },
                    "error": {
                        "type": "string",
                        "title": "Error"
                    },
                    "message": {
                        "type": "string",
                        "title": "Message"
                    },
                    "timestamp": {
                        "type": "string",
                        "title": "Timestamp"
                    },
                    "correlation_id": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Correlation Id",
                        "description": "Per-invocation UUID4 hex, present on replay error responses (501 / 409 / 502) so operators can grep logs by it. Omitted for non-replay error paths."
                    }
                },
                "type": "object",
                "required": [
                    "error",
                    "message",
                    "timestamp"
                ],
                "title": "RealtimeAlertErrorResponse",
                "description": "Generic error envelope.",
                "example": {
                    "error": "validation_error",
                    "message": "live_stream_url must be an RTSP URL (rtsp://...)",
                    "status": "error",
                    "timestamp": "2025-06-01T12:00:00Z"
                }
            },
...
        }
    }
}